Ajout: -Finalisation de contraintes de "Join" avec etat evenement

+++ src/Controller/EventController.php
@@ -55,19 +55,27 @@ class EventController extends AbstractController
     #[Route('/add', name: 'create')]
 public function create(
                  EntityManagerInterface $entityManager,
-                 Request $request
+                 Request $request,
+                 UserRepository $userRepository,
+                 StateEventRepository $stateEventRepository,
+
     ):Response
     {
+        $statesEvent = $stateEventRepository->findAll();
+        $stateEventCreated=$statesEvent[0]; // Sortie Etat "Created"
+        $user = $userRepository->find($this->getUser()->getId());
         $event = new Event();
         $eventForm = $this->createForm(EventType::class, $event);
         $eventForm->handleRequest($request);
         if($eventForm->isSubmitted() && $eventForm->isValid()){
             $event->setOrganizer($this->getUser());
+            $event->addParticipant($user);
+            $event->setStateEvent($stateEventCreated);
 
             $entityManager->persist($event);
             $entityManager->flush();
 
-            $this->addFlash('success', "evenement ajoutée! -_*");
+            $this->addFlash('success', "Sortie ajoutée !");
             return $this->redirectToRoute('event_list');
         }
 
@@ -86,10 +94,10 @@ public function create(
     {
         $event = $eventRepository->find($id);
         if(!$event){
-            throw $this->createNotFoundException("Cette événement n'a pas été trouvé");
+            throw $this->createNotFoundException("La sortie n'a pas été trouvé");
         }
         if($event->getOrganizer() !== $this->getUser()){
-            throw $this->createNotFoundException("Vous n'êtes pas l'organisateur, vous ne pouvez pas modifier cet évènement");
+            throw $this->createNotFoundException("Vous n'êtes pas l'organisateur, vous ne pouvez pas modifier cette sortie");
         }
 
         $eventForm = $this->createForm(EventType::class, $event);
@@ -98,7 +106,7 @@ public function create(
             $entityManager->persist($event);
             $entityManager->flush();
 
-            $this->addFlash('success', "événement modifiée");
+            $this->addFlash('success', "sortie modifiée");
             return $this->redirectToRoute('event_list');
         }
         return $this->render('event/updateEvent.html.twig', [
@@ -116,12 +124,15 @@ public function create(
     {
         $event = $eventRepository->find($id);
         if(!$event){
-            throw $this->createNotFoundException("Cette événement n'a pas été trouvée");
+            throw $this->createNotFoundException("Cette sortie n'a pas été trouvée");
+        }
+        if($event->getOrganizer() !== $this->getUser()){
+            throw $this->createNotFoundException("Vous n'êtes pas l'organisateur, vous ne pouvez pas supprimer cette sortie");
         }
         $entityManager->remove($event);
         $entityManager->flush();
 
-        $this->addFlash( 'success' ,'événement supprimée !!');
+        $this->addFlash( 'success' ,'Sortie supprimée !!');
         return $this->redirectToRoute('event_list');
     }
     #[Route('/join/{id}', name: 'join')]
@@ -129,20 +140,33 @@ public function create(
         EntityManagerInterface $entityManager,
         EventRepository $eventRepository,
         UserRepository $userRepository,
+        StateEventRepository $stateEventRepository,
         Request $request,
         int $id
     ):Response
     {
-        $user = $userRepository->find($this->getUser()->getId());
+        $statesEvent = $stateEventRepository->findAll();
+        $stateEventOpen=$statesEvent[1]; // Evenement "Open"
+        $user = $userRepository->find($this->getUser()->getId()); // Utilisateur connecté
         $event = $eventRepository->find($id);
         if(!$event){
-            throw $this->createNotFoundException("Cette événement n'a pas été trouvé");
+            throw $this->createNotFoundException("Cette sortie n'a pas été trouvée");
         }
         if($event->getOrganizer() == $this->getUser()){
-            throw $this->createNotFoundException("Vous êtes l'organisateur, vous participez déjà à cet évènement");
+            throw $this->createNotFoundException("Vous êtes l'organisateur, vous participez déjà à cette sortie");
         }
+        if($event->getParticipants()->contains($user)){
+            throw $this->createNotFoundException("Vous êtes déjà inscrit à cette sortie");
+        }
+
         if($event->getParticipants()->count() == $event->getMaxParticipants()){
-            throw $this->createNotFoundException("Cet évènement est complet");
+            throw $this->createNotFoundException("Cette sortie est complète");
+        }
+        if($event->getDateLine() < new \DateTime('now')){
+            throw $this->createNotFoundException("Cette sortie est déjà terminée");
+        }
+        if($event->getStateEvent() !== $stateEventOpen){
+            throw $this->createNotFoundException("Les inscriptions pour cette sortie ne sont pas ouvertes");
         }
 
         $event->addParticipant($user);
